pyproject.toml
[tool.poetry]
name = "photographyanonymizator"
version = "0.1.0"
description = ""
authors = ["Rubén Chuliá Mena <rubchume@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-vision = "^2.3.1"
opencv-python = "^4.5.1"
django = "^3.2"
numpy = "^1.20.2"
djangorestframework = "^3.12.4"
Pillow = "^8.2.0"
django-background-tasks = "^1.2.5"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
nox = "^2020.5.24"
safety = "^1.9.0"
flake8 = "^3.8.3"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
black = "^19.10b0"
flake8-black = "^0.2.1"
pytest-cov = "^2.10.1"
coverage = {extras = ["toml"], version = "^5.2.1"}
mypy = "^0.790"
typeguard = "^2.10.0"
pytest-mock = "^3.3.1"
requests-mock = "^1.8.0"
importlib-metadata = "^4.0.1"
pytest-django = "^4.2.0"
beautifulsoup4 = "^4.9.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["src", "core"]

[tool.coverage.report]
show_missing = true
fail_under = 100
